Update utils.py
Took out powerset() and generate_unique_expressions_of_len() as they are not used. Changed variable name "nr" into "bootrap_id" to be more informative and in line with the other files. Changed element_wise_binary_entropy() so it does not warn for division by zero (as that case is being dealt with appropriately).