Write new README file to accompany the code of CogSci2021 paper
The README.org file accompanied the original version of the forked code. Replaced it with a README.md file that gives some general information about the paper and which programs to use for generating and analysing data. Kept some lines from README.org that are still relevant for this version of the code.